+++ app/components/sheets/AlertDetail.tsx
@@ -1,7 +1,7 @@
 import { BottomSheetModal, BottomSheetScrollView, BottomSheetView} from "@gorhom/bottom-sheet";
 import { View, TouchableOpacity, useWindowDimensions} from "react-native";
 import Ionicons from '@expo/vector-icons/Ionicons';
-import useAppStore from "../../stores/useAppStore";
+import useAppStore from "../../data/app_state";
 import SheetHeader from "../ui/SheetHeader";
 import RenderHtml from 'react-native-render-html';
