+++ app/components/ui/DateSelector.tsx
@@ -1,5 +1,5 @@
 import { Ionicons } from '@expo/vector-icons';
-import useAppStore from '../../stores/useAppStore';
+import useAppStore from '../../data/app_state';
 import React from 'react';
 import { View, Text, TouchableOpacity } from 'react-native';
