+++ app/components/ui/IconPill.tsx
@@ -1,4 +1,4 @@
-import useAppStore from '../../stores/useAppStore'
+import useAppStore from '../../data/app_state'
 import React, { memo } from 'react'
 import { View, Text, ViewProps } from 'react-native'
